Update sliders.js
This update get's rid of the jump at the start of the slide. On almost every slider know to man there is a 15px jump. This one has it to. If you log out deltaX you'll see it almost always starts with -16px. That's a noticeable jump. This change fixes it so that the slide starts from 0 and smoothly goes from there.