Filter children to slides only when determining number of slides
When using MVC frameworks (Ember) to generate the HTML, additional
`<script>` nodes are inserted inside the `.slide-group` element. These
extra nodes get returned by `slider.children`, causing an incorrect
number of slides to be calculated, and hence when scrolling to the end
of the slides, you scroll into nothingness.

This patch counts only the `.slide` nodes within the slider to get the
correct slide count.